deleting alert

+++ components/FavoriteComponents.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import {View, FlatList, Text } from 'react-native';
+import {View, FlatList, Text, Alert } from 'react-native';
 import {ListItem} from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
@@ -28,13 +28,31 @@ class Favorites extends Component {
         const { navigate } = this.props.navigation;
 
         const renderMenuItem = ({item, index}) => {
-            const rightButton= [
-            {
-                text: 'Delete',
-                type: 'delete',
-                onPress: () => this.props.deleteFavorite(item.id)
-            }
-        ];
+            const rightButton = [
+                {
+                    text: 'Delete', 
+                    type: 'delete',
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                        
+                    }
+                }
+            ];
 
             return(
                 <Swipeout right={rightButton} autoClose={true}>
